+++ db/migrate/20181228181118_create_poste_depenses.rb
@@ -2,7 +2,7 @@ class CreatePosteDepenses < ActiveRecord::Migration[5.2]
   def change
     create_table :poste_depenses do |t|
       t.string :categorie
-      t.integer :montant
+      t.float :montant
       t.references :budget, foreign_key: true
 
       t.timestamps
